third_party/blink/web_tests/platform/generic/fast/dom/Document/scrollingElement-noLayout-quirks-expected.txt

Verify that document.scrollingElement updates style in quirks mode without triggering layout

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.needsLayoutCount() is 0
PASS document.scrollingElement is document.body
Verifying layout hasn't been triggered
PASS internals.needsLayoutCount() is 3
But style update was
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
Verify style change is reflected - body is now a real scrolling block element
PASS document.scrollingElement is null
Verifying layout still hasn't been triggered
PASS internals.needsLayoutCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
